requirements.txt
py-cord
PyNaCl
aiodns
brotlipy
msgspec
python-dateutil